pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "polygon-mcp-server"
version = "1.0.0"
description = "Advanced Polygon.io MCP Server with Options Analytics, Greeks, and Real-time Market Data"
authors = [
    {name = "Polygon MCP Team", email = "support@polygon.io"},
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "anthropic>=0.25.0",
    "httpx>=0.24.0",
    "pydantic>=2.0.0",
    "scipy>=1.9.0",
    "numpy>=1.21.0",
    "mcp>=1.0.0",
    "polygon-api-client>=1.14.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
]

[project.urls]
Homepage = "https://github.com/polygon-io/mcp-server"
Repository = "https://github.com/polygon-io/mcp-server"

[project.scripts]
polygon-mcp-server = "polygon_mcp_server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/polygon_mcp_server"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/README.md",
    "/LICENSE",
] 